+++ src/components/Header/Header.jsx
@@ -1,12 +1,12 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useState } from 'react';
 
 import { Icon } from '../Icon/Icon';
-import { Sidebar } from '../Sidebar/Sidebar';
 import { Loader } from '../Loader/Loader';
 import { ThemeSwitcher } from './ThemeSwitcher/ThemeSwitcher';
 import { UserProfile } from './User/UserProfile';
 
-import { SButton_button, SHeader_header, SDiv_Backdrop } from './Header.styled';
+import { SButton_button, SHeader_header } from './Header.styled';
+import { SidebarModal } from './SidebarModal/sidebarModal';
 
 export const Header = () => {
   const [isOpenSidebar, setIsOpenSidebar] = useState(false);
@@ -22,24 +22,13 @@ export const Header = () => {
     }
   };
 
-  useEffect(() => {
-    const handleKeydown = (event) => {
-      const { code } = event;
-      if (code === 'Escape') toggleSidebar();
-    };
-    document.addEventListener('keydown', handleKeydown);
-
-    return () => {
-      document.removeEventListener('keydown', handleKeydown);
-    };
-  }, [toggleSidebar]);
-
   return (
     <SHeader_header>
       {isOpenSidebar && (
-        <SDiv_Backdrop onClick={handleBackdropClick}>
-          <Sidebar className="sidebarHeader" />
-        </SDiv_Backdrop>
+        <SidebarModal
+          handleBackdropClick={handleBackdropClick}
+          toggleSidebar={toggleSidebar}
+        />
       )}
       <SButton_button onClick={toggleSidebar}>
         <Icon id={'menu'} />
