+++ src/components/Header/Header.styled.js
@@ -37,18 +37,3 @@ export const SButton_button = styled.button`
     display: none;
   }
 `;
-
-export const SDiv_Backdrop = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100vh;
-  background-color: rgba(21, 21, 21, 0.5);
-  backdrop-filter: blur(5px);
-  z-index: 999;
-
-  @media only screen and (min-width: 1440px) {
-    display: none;
-  }
-`;
