+++ src/components/Header/SidebarModal/SidebarModal.styled.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const SDiv_Backdrop = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100vh;
+  background-color: rgba(21, 21, 21, 0.5);
+  backdrop-filter: blur(5px);
+  z-index: 999;
+
+  @media only screen and (min-width: 1440px) {
+    display: none;
+  }
+`;
